refactor: core tag translators now are the tag definition
