+++ package.json
@@ -3,7 +3,7 @@
   "devDependencies": {
     "@commitlint/cli": "^16.0.2",
     "@commitlint/config-conventional": "^16.0.0",
-    "@marko/compiler": "^5.17.10",
+    "@marko/compiler": "^5.19.1",
     "@rollup/plugin-replace": "3.0.1",
     "@types/jsdom": "^16.2.14",
     "@types/mocha": "^9.0.0",
