+++ packages/translator/package.json
@@ -4,7 +4,7 @@
   "version": "0.0.1",
   "bugs": "https://github.com/marko-js/x/issues/new?template=Bug_report.md",
   "dependencies": {
-    "@marko/babel-utils": "5.17.4",
+    "@marko/babel-utils": "^5.19.1",
     "@marko/runtime-fluurt": "^0.0.1",
     "tslib": "^2.3.1"
   },
