+++ packages/translator/src/core/attrs.ts
@@ -1,0 +1,15 @@
+import type { Tag } from "@marko/babel-utils";
+
+export default {
+  translate(tag) {
+    tag.remove();
+  },
+  attributes: {},
+  autocomplete: [
+    {
+      displayText: "attrs/{ ... }",
+      description: "Use to receive the attributes passed into this template.",
+      snippet: "attrs/{ $1 }$2",
+    },
+  ],
+} as Tag;
