--- packages/translator/src/core/translate-attrs.ts
@@ -1,5 +1,0 @@
-import type { types as t } from "@marko/compiler";
-
-export default function enter(tag: t.NodePath<t.MarkoTag>) {
-  tag.remove();
-}
